Fix CI & tests
Needed to say it can run on linux, as well as Windows (where the
main dev work is being done) so that it can run Github Actions -
and work on other linux machines.

There's a Lint/ScriptPermission problem showing up on Github Actions
that I don't get locally, because of the different OS'es. It's not
easily fixable on Windows and not important anyway, so let's ignore
it.

Finally I forgot to skip a test in the test suite (that requires
real username/password to work).